Made some bug-fixes and optimizations that I found while working with unrelated things:   - Fixed a calculation error in MediaMonitor   - Fixed a NPE in RarredEntry   - Made PlayerSort in PlayerFactory Serializable since it implements Comparator   - Made ISO extensions immutable   - Made PLAYLIST extensions immutable   - Fixed a calculation error in HelpTab   - Made a small performance improvement in TracesTab   - Narrowed Exception handling in BasicPlayer   - Made a small performance improvement in CoverArtArchiveUtil and CoverUtil   - Fixed a NPE in FreedesktopTrash

+++ src/main/java/net/pms/dlna/MediaMonitor.java
@@ -151,9 +151,9 @@ public class MediaMonitor extends VirtualFolder {
 		 */
 		double elapsed;
 		if (res.getLastStartPosition() == 0) {
-			elapsed = (System.currentTimeMillis() - res.getStartTime()) / 1000;
+			elapsed = (double) (System.currentTimeMillis() - res.getStartTime()) / 1000;
 		} else {
-			elapsed = (System.currentTimeMillis() - res.getLastStartSystemTime()) / 1000;
+			elapsed = (double) (System.currentTimeMillis() - res.getLastStartSystemTime()) / 1000;
 			elapsed += res.getLastStartPosition();
 		}
